added NodeBuilderFactory and threaded through chunker
