+++ go/store/prolly/map_test.go
@@ -89,6 +89,15 @@ func TestMap(t *testing.T) {
 	}
 }
 
+func TestNewEmptyNode(t *testing.T) {
+	empty := newEmptyNode(sharedPool)
+	assert.Equal(t, 0, empty.Level())
+	assert.Equal(t, 0, empty.Count())
+	assert.Equal(t, 68, empty.Size())
+	assert.Equal(t, 0, empty.TreeCount())
+	assert.True(t, empty.IsLeaf())
+}
+
 func makeProllyMap(t *testing.T, count int) (testMap, [][2]val.Tuple) {
 	kd := val.NewTupleDescriptor(
 		val.Type{Enc: val.Uint32Enc, Nullable: false},
@@ -122,7 +131,7 @@ func prollyMapFromTuples(t *testing.T, kd, vd val.TupleDesc, tuples [][2]val.Tup
 	ctx := context.Background()
 	ns := tree.NewTestNodeStore()
 
-	chunker, err := tree.NewEmptyChunker(ctx, ns)
+	chunker, err := tree.NewEmptyChunker(ctx, ns, newNodeBuilder)
 	require.NoError(t, err)
 
 	for _, pair := range tuples {
