+++ go/store/prolly/ordered_tree.go
@@ -69,9 +69,8 @@ func mergeOrderedTrees[K, V ~[]byte, O ordering[K]](
 	l, r, base orderedTree[K, V, O],
 	cb tree.CollisionFn,
 ) (orderedTree[K, V, O], error) {
-	sfn, cfn := base.searchNode, base.compareItems
-
-	root, err := tree.ThreeWayMerge(ctx, base.ns, l.root, r.root, base.root, sfn, cfn, cb)
+	cfn, fact := base.compareItems, newNodeBuilder
+	root, err := tree.ThreeWayMerge(ctx, base.ns, l.root, r.root, base.root, cfn, cb, fact)
 	if err != nil {
 		return orderedTree[K, V, O]{}, err
 	}
