+++ go/store/prolly/tree/node_cursor_test.go
@@ -53,7 +53,7 @@ func testNewCursorAtItem(t *testing.T, count int) {
 func randomTree(t *testing.T, count int) (Node, [][2]Item, NodeStore) {
 	ctx := context.Background()
 	ns := NewTestNodeStore()
-	chkr, err := newEmptyChunker(ctx, ns)
+	chkr, err := newEmptyChunker(ctx, ns, newTestBuilder)
 	require.NoError(t, err)
 
 	items := randomTupleItemPairs(count / 2)
