+++ go/store/prolly/tree/testutils.go
@@ -15,20 +15,17 @@
 package tree
 
 import (
-	"context"
+	"encoding/binary"
 	"fmt"
 	"math"
 	"math/rand"
 	"sort"
-	"strings"
-	"testing"
 
-	"github.com/stretchr/testify/require"
-	"gonum.org/v1/plot"
-	"gonum.org/v1/plot/plotter"
-	"gonum.org/v1/plot/vg"
+	fb "github.com/google/flatbuffers/go"
 
+	"github.com/dolthub/dolt/go/gen/fb/serial"
 	"github.com/dolthub/dolt/go/store/chunks"
+	"github.com/dolthub/dolt/go/store/pool"
 	"github.com/dolthub/dolt/go/store/val"
 )
 
@@ -49,15 +46,6 @@ func NewTupleLeafNode(keys, values []val.Tuple) Node {
 		vs[i] = Item(values[i])
 	}
 	return newLeafNode(ks, vs)
-}
-
-func newLeafNode(keys, values []Item) Node {
-	b := &nodeBuilder{
-		keys:   keys,
-		values: values,
-		level:  0,
-	}
-	return b.Build(sharedPool)
 }
 
 func RandomTuplePairs(count int, keyDesc, valDesc val.TupleDesc) (items [][2]val.Tuple) {
@@ -142,48 +130,6 @@ func AscendingUintTuples(count int) (tuples [][2]val.Tuple, desc val.TupleDesc) 
 	return
 }
 
-func AscendingIntTuples(t *testing.T, count int) (tuples [][2]val.Tuple, desc val.TupleDesc) {
-	desc = val.NewTupleDescriptor(val.Type{Enc: val.Int32Enc})
-	bld := val.NewTupleBuilder(desc)
-	tuples = make([][2]val.Tuple, count)
-	for i := range tuples {
-		bld.PutInt32(0, int32(i))
-		tuples[i][0] = bld.Build(sharedPool)
-		bld.PutInt32(0, int32(i+count))
-		tuples[i][1] = bld.Build(sharedPool)
-	}
-	return
-}
-
-// Map<Tuple<Uint32,Uint32>,Tuple<Uint32,Uint32>>
-func AscendingCompositeIntTuples(count int) (keys, values []val.Tuple, desc val.TupleDesc) {
-	desc = val.NewTupleDescriptor(val.Type{Enc: val.Uint32Enc}, val.Type{Enc: val.Uint32Enc})
-	bld := val.NewTupleBuilder(desc)
-
-	tups := make([]val.Tuple, count*2)
-	for i := range tups {
-		bld.PutUint32(0, uint32(i))
-		bld.PutUint32(1, uint32(i))
-		tups[i] = bld.Build(sharedPool)
-	}
-	keys, values = tups[:count], tups[count:]
-	return
-}
-
-// assumes a sorted list
-func deduplicateTuples(desc val.TupleDesc, tups [][2]val.Tuple) (uniq [][2]val.Tuple) {
-	uniq = make([][2]val.Tuple, 1, len(tups))
-	uniq[0] = tups[0]
-
-	for i := 1; i < len(tups); i++ {
-		cmp := desc.Compare(tups[i-1][0], tups[i][0])
-		if cmp < 0 {
-			uniq = append(uniq, tups[i])
-		}
-	}
-	return
-}
-
 func RandomTuple(tb *val.TupleBuilder) (tup val.Tuple) {
 	for i, typ := range tb.Desc.Types {
 		randomField(tb, i, typ)
@@ -209,6 +155,29 @@ func ShuffleTuplePairs(items [][2]val.Tuple) {
 	testRand.Shuffle(len(items), func(i, j int) {
 		items[i], items[j] = items[j], items[i]
 	})
+}
+
+func newLeafNode(keys, values []Item) Node {
+	b := &testBuilder{
+		keys:   keys,
+		values: values,
+		level:  0,
+	}
+	return b.Build(sharedPool)
+}
+
+// assumes a sorted list
+func deduplicateTuples(desc val.TupleDesc, tups [][2]val.Tuple) (uniq [][2]val.Tuple) {
+	uniq = make([][2]val.Tuple, 1, len(tups))
+	uniq[0] = tups[0]
+
+	for i := 1; i < len(tups); i++ {
+		cmp := desc.Compare(tups[i-1][0], tups[i][0])
+		if cmp < 0 {
+			uniq = append(uniq, tups[i])
+		}
+	}
+	return
 }
 
 func randomField(tb *val.TupleBuilder, idx int, typ val.Type) {
@@ -261,123 +230,157 @@ func randomField(tb *val.TupleBuilder, idx int, typ val.Type) {
 	}
 }
 
-func fmtTupleList(tuples [][2]val.Tuple, kd, vd val.TupleDesc) string {
-	var sb strings.Builder
-	sb.WriteString("{ ")
-	for _, kv := range tuples {
-		if kv[0] == nil || kv[1] == nil {
-			break
-		}
-		sb.WriteString(kd.Format(kv[0]))
-		sb.WriteString(": ")
-		sb.WriteString(vd.Format(kv[1]))
-		sb.WriteString(", ")
+func newTestBuilder(level int) *testBuilder {
+	return &testBuilder{level: level}
+}
+
+var _ NodeBuilderFactory[*testBuilder] = newTestBuilder
+
+type testBuilder struct {
+	keys, values []Item
+	size, level  int
+
+	subtrees SubtreeCounts
+}
+
+var _ NodeBuilder = &testBuilder{}
+
+func (tb *testBuilder) StartNode() {
+	tb.reset()
+}
+
+func (tb *testBuilder) HasCapacity(key, value Item) bool {
+	sum := tb.size + len(key) + len(value)
+	return sum <= int(MaxVectorOffset)
+}
+
+func (tb *testBuilder) AddItems(key, value Item, subtree uint64) {
+	tb.keys = append(tb.keys, key)
+	tb.values = append(tb.values, value)
+	tb.size += len(key) + len(value)
+	tb.subtrees = append(tb.subtrees, subtree)
+}
+
+func (tb *testBuilder) Count() int {
+	return len(tb.keys)
+}
+
+func (tb *testBuilder) Build(pool pool.BuffPool) (node Node) {
+	var (
+		keyTups, keyOffs fb.UOffsetT
+		valTups, valOffs fb.UOffsetT
+		refArr, cardArr  fb.UOffsetT
+	)
+
+	keySz, valSz, bufSz := measureNodeSize(tb.keys, tb.values, tb.subtrees)
+	b := getMapBuilder(pool, bufSz)
+
+	// serialize keys and offsets
+	keyTups = writeItemBytes(b, tb.keys, keySz)
+	serial.ProllyTreeNodeStartKeyOffsetsVector(b, len(tb.keys)-1)
+	keyOffs = b.EndVector(writeItemOffsets(b, tb.keys, keySz))
+
+	if tb.level == 0 {
+		// serialize value tuples for leaf nodes
+		valTups = writeItemBytes(b, tb.values, valSz)
+		serial.ProllyTreeNodeStartValueOffsetsVector(b, len(tb.values)-1)
+		valOffs = b.EndVector(writeItemOffsets(b, tb.values, valSz))
+	} else {
+		// serialize child refs and subtree counts for internal nodes
+		refArr = writeItemBytes(b, tb.values, valSz)
+		cardArr = writeCountArray(b, tb.subtrees)
 	}
-	sb.WriteString("}")
-	return sb.String()
+
+	// populate the node's vtable
+	serial.ProllyTreeNodeStart(b)
+	serial.ProllyTreeNodeAddKeyItems(b, keyTups)
+	serial.ProllyTreeNodeAddKeyOffsets(b, keyOffs)
+	if tb.level == 0 {
+		serial.ProllyTreeNodeAddValueItems(b, valTups)
+		serial.ProllyTreeNodeAddValueOffsets(b, valOffs)
+		serial.ProllyTreeNodeAddTreeCount(b, uint64(len(tb.keys)))
+	} else {
+		serial.ProllyTreeNodeAddAddressArray(b, refArr)
+		serial.ProllyTreeNodeAddSubtreeCounts(b, cardArr)
+		serial.ProllyTreeNodeAddTreeCount(b, tb.subtrees.Sum())
+	}
+	serial.ProllyTreeNodeAddKeyType(b, serial.ItemTypeTupleFormatAlpha)
+	serial.ProllyTreeNodeAddValueType(b, serial.ItemTypeTupleFormatAlpha)
+	serial.ProllyTreeNodeAddTreeLevel(b, uint8(tb.level))
+	b.Finish(serial.ProllyTreeNodeEnd(b))
+	tb.reset()
+
+	buf := b.FinishedBytes()
+	return NodeFromBytes(buf)
 }
 
-type Samples []int
-
-func (s Samples) Summary() string {
-	f := "mean: %8.2f \t stddev: %8.2f \t p50: %5d \t p90: %5d \t p99: %5d \t p99.9: %5d \t p100: %5d"
-	p50, p90, p99, p999, p100 := s.percentiles()
-	return fmt.Sprintf(f, s.mean(), s.stdDev(), p50, p90, p99, p999, p100)
+func (tb *testBuilder) reset() {
+	// buffers are copied, it's safe to re-use the memory.
+	tb.keys = tb.keys[:0]
+	tb.values = tb.values[:0]
+	tb.size = 0
+	tb.subtrees = tb.subtrees[:0]
 }
 
-func (s Samples) count() float64 {
-	return float64(len(s))
+func getMapBuilder(pool pool.BuffPool, sz int) (b *fb.Builder) {
+	b = fb.NewBuilder(0)
+	buf := pool.Get(uint64(sz))
+	b.Bytes = buf[:0]
+	return
 }
 
-func (s Samples) sum() (total float64) {
-	for _, v := range s {
-		total += float64(v)
+// measureNodeSize returns the exact Size of the tuple vectors for keys and values,
+// and an estimate of the overall Size of the final flatbuffer.
+func measureNodeSize(keys, values []Item, subtrees []uint64) (keySz, valSz, bufSz int) {
+	for i := range keys {
+		keySz += len(keys[i])
+		valSz += len(values[i])
+	}
+	refCntSz := len(subtrees) * binary.MaxVarintLen64
+
+	// constraints enforced upstream
+	if keySz > int(MaxVectorOffset) {
+		panic(fmt.Sprintf("key vector exceeds Size limit ( %d > %d )", keySz, MaxVectorOffset))
+	}
+	if valSz > int(MaxVectorOffset) {
+		panic(fmt.Sprintf("value vector exceeds Size limit ( %d > %d )", valSz, MaxVectorOffset))
+	}
+
+	bufSz += keySz + valSz               // tuples
+	bufSz += refCntSz                    // subtree counts
+	bufSz += len(keys)*2 + len(values)*2 // offsets
+	bufSz += 8 + 1 + 1 + 1               // metadata
+	bufSz += 72                          // vtable (approx)
+
+	return
+}
+
+func writeItemBytes(b *fb.Builder, items []Item, sumSz int) fb.UOffsetT {
+	b.Prep(fb.SizeUOffsetT, sumSz)
+
+	stop := int(b.Head())
+	start := stop - sumSz
+	for _, item := range items {
+		copy(b.Bytes[start:stop], item)
+		start += len(item)
+	}
+
+	start = stop - sumSz
+	return b.CreateByteVector(b.Bytes[start:stop])
+}
+
+func writeItemOffsets(b *fb.Builder, items []Item, sz int) (cnt int) {
+	off := sz
+	for i := len(items) - 1; i > 0; i-- { // omit first offset
+		off -= len(items[i])
+		b.PrependUint16(uint16(off))
+		cnt++
 	}
 	return
 }
 
-func (s Samples) mean() float64 {
-	return s.sum() / float64(len(s))
-}
-
-func (s Samples) stdDev() float64 {
-	var acc float64
-	u := s.mean()
-	for _, v := range s {
-		d := float64(v) - u
-		acc += d * d
-	}
-	return math.Sqrt(acc / s.count())
-}
-
-func (s Samples) percentiles() (p50, p90, p99, p999, p100 int) {
-	sort.Ints(s)
-	l := len(s)
-	p50 = s[l/2]
-	p90 = s[(l*9)/10]
-	p99 = s[(l*99)/100]
-	p999 = s[(l*999)/1000]
-	p100 = s[l-1]
-	return
-}
-
-func PrintTreeSummaryByLevel(t *testing.T, nd Node, ns NodeStore) {
-	ctx := context.Background()
-
-	sizeByLevel := make([]Samples, nd.Level()+1)
-	cardByLevel := make([]Samples, nd.Level()+1)
-	err := WalkNodes(ctx, nd, ns, func(ctx context.Context, nd Node) error {
-		lvl := nd.Level()
-		sizeByLevel[lvl] = append(sizeByLevel[lvl], nd.Size())
-		cardByLevel[lvl] = append(cardByLevel[lvl], int(nd.count))
-		return nil
-	})
-	require.NoError(t, err)
-
-	fmt.Println("pre-edit map Summary: ")
-	fmt.Println("| Level | count | avg Size \t  p50 \t  p90 \t p100 | avg card \t  p50 \t  p90 \t p100 |")
-	for i := nd.Level(); i >= 0; i-- {
-		sizes, cards := sizeByLevel[i], cardByLevel[i]
-		sp50, _, sp90, _, sp100 := sizes.percentiles()
-		cp50, _, cp90, _, cp100 := cards.percentiles()
-		fmt.Printf("| %5d | %5d | %8.2f \t %4d \t %4d \t %4d | %8.2f \t %4d \t %4d \t %4d |\n",
-			i, len(cards), sizes.mean(), sp50, sp90, sp100, cards.mean(), cp50, cp90, cp100)
-	}
-	fmt.Println()
-}
-
-func plotNodeSizeDistribution(t *testing.T, name string, nd Node, ns NodeStore) {
-	data, err := measureTreeNodes(nd, ns)
-	require.NoError(t, err)
-	plotIntHistogram(name, data)
-}
-
-func measureTreeNodes(nd Node, ns NodeStore) (Samples, error) {
-	ctx := context.Background()
-	data := make(Samples, 0, 1024)
-	err := WalkNodes(ctx, nd, ns, func(ctx context.Context, nd Node) error {
-		data = append(data, nd.Size())
-		return nil
-	})
-	return data, err
-}
-
-func plotIntHistogram(name string, data []int) {
-	var values plotter.Values
-	for _, d := range data {
-		values = append(values, float64(d))
-	}
-
-	p := plot.New()
-	p.Title.Text = "histogram plot"
-
-	hist, err := plotter.NewHist(values, 50)
-	if err != nil {
-		panic(err)
-	}
-	p.Add(hist)
-
-	if err := p.Save(3*vg.Inch, 3*vg.Inch, name); err != nil {
-		panic(err)
-	}
+func writeCountArray(b *fb.Builder, sc SubtreeCounts) fb.UOffsetT {
+	// todo(andy) write without copy
+	arr := WriteSubtreeCounts(sc)
+	return b.CreateByteVector(arr)
 }
